# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatOTP - WhatsApp OTP Verification SaaS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  suspended?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserWithSubscriptionRole =
  (typeof UserWithSubscriptionRole)[keyof typeof UserWithSubscriptionRole];

export const UserWithSubscriptionRole = {
  user: "user",
  admin: "admin",
} as const;

export type PlanPeriod = (typeof PlanPeriod)[keyof typeof PlanPeriod];

export const PlanPeriod = {
  weekly: "weekly",
  monthly: "monthly",
  yearly: "yearly",
} as const;

export interface Plan {
  id: number;
  name: string;
  period: PlanPeriod;
  price: number;
  currency: string;
  otpLimit: number;
  features: string[];
  popular?: boolean;
  allowCustomNumber?: boolean;
  allowUsaNumbers?: boolean;
}

export type SubscriptionStatus =
  (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus];

export const SubscriptionStatus = {
  active: "active",
  cancelled: "cancelled",
  expired: "expired",
} as const;

export interface Subscription {
  id: number;
  userId: number;
  planId: number;
  plan?: Plan;
  status: SubscriptionStatus;
  startDate: string;
  endDate: string;
  otpUsed?: number;
}

export interface UserWithSubscription {
  id: number;
  email: string;
  name: string;
  role: UserWithSubscriptionRole;
  suspended?: boolean;
  createdAt: string;
  subscription?: Subscription;
  otpCount?: number;
}

export interface SubscriptionInput {
  planId: number;
}

export type ApiKeyStatus = (typeof ApiKeyStatus)[keyof typeof ApiKeyStatus];

export const ApiKeyStatus = {
  active: "active",
  revoked: "revoked",
} as const;

export interface ApiKey {
  id: number;
  name: string;
  keyPrefix: string;
  status: ApiKeyStatus;
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
  requestCount?: number;
}

export type ApiKeyWithSecretStatus =
  (typeof ApiKeyWithSecretStatus)[keyof typeof ApiKeyWithSecretStatus];

export const ApiKeyWithSecretStatus = {
  active: "active",
  revoked: "revoked",
} as const;

export interface ApiKeyWithSecret {
  id: number;
  name: string;
  keyPrefix: string;
  status: ApiKeyWithSecretStatus;
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
  requestCount?: number;
  secret: string;
}

export interface ApiKeyInput {
  /** @minLength 1 */
  name: string;
}

export interface AssignPlanInput {
  planId: number;
}

export interface DisplayNameInput {
  /**
   * @minLength 1
   * @maxLength 25
   */
  displayName: string;
}

export interface PaymentInitInput {
  planId: number;
}

export interface PaymentInitResponse {
  authorizationUrl: string;
  reference: string;
  planId: number;
  planName?: string;
  amount: number;
  currency: string;
}

export interface UserWhatsappNumberInput {
  /** @minLength 7 */
  phoneNumber: string;
  label?: string;
}

export type WhatsappNumberCountry =
  (typeof WhatsappNumberCountry)[keyof typeof WhatsappNumberCountry];

export const WhatsappNumberCountry = {
  us: "us",
  ng: "ng",
  uk: "uk",
} as const;

export type WhatsappNumberStatus =
  (typeof WhatsappNumberStatus)[keyof typeof WhatsappNumberStatus];

export const WhatsappNumberStatus = {
  disconnected: "disconnected",
  connecting: "connecting",
  connected: "connected",
  error: "error",
} as const;

export interface WhatsappNumber {
  id: number;
  phoneNumber: string;
  country: WhatsappNumberCountry;
  /** @nullable */
  label?: string | null;
  status: WhatsappNumberStatus;
  sessionActive?: boolean;
  otpSentCount?: number;
  createdAt: string;
}

export type WhatsappNumberInputCountry =
  (typeof WhatsappNumberInputCountry)[keyof typeof WhatsappNumberInputCountry];

export const WhatsappNumberInputCountry = {
  us: "us",
  ng: "ng",
  uk: "uk",
} as const;

export interface WhatsappNumberInput {
  /** @minLength 7 */
  phoneNumber: string;
  country: WhatsappNumberInputCountry;
  label?: string;
}

export type QrCodeResponseStatus =
  (typeof QrCodeResponseStatus)[keyof typeof QrCodeResponseStatus];

export const QrCodeResponseStatus = {
  disconnected: "disconnected",
  connecting: "connecting",
  connected: "connected",
  error: "error",
} as const;

export interface QrCodeResponse {
  /** @nullable */
  qrCode: string | null;
  status: QrCodeResponseStatus;
}

export interface OtpSendInput {
  /** @minLength 7 */
  phoneNumber: string;
  template?: string;
}

export interface OtpSendResponse {
  requestId: string;
  expiresAt: string;
  message?: string;
}

export interface OtpVerifyInput {
  requestId: string;
  /** @minLength 4 */
  code: string;
}

export interface OtpVerifyResponse {
  valid: boolean;
  message?: string;
}

export type OtpLogStatus = (typeof OtpLogStatus)[keyof typeof OtpLogStatus];

export const OtpLogStatus = {
  pending: "pending",
  sent: "sent",
  verified: "verified",
  failed: "failed",
  expired: "expired",
} as const;

export interface OtpLog {
  id: number;
  phoneNumber: string;
  status: OtpLogStatus;
  requestId?: string;
  /** @nullable */
  country?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  otpSentToday: number;
  otpSentMonth: number;
  otpSuccessRate: number;
  remainingOtps: number;
  activeApiKeys: number;
  recentActivity?: OtpLog[];
}

export type AdminStatsOtpByCountryItem = {
  country: string;
  count: number;
};

export interface AdminStats {
  totalUsers: number;
  activeSubscriptions: number;
  totalOtpSent: number;
  connectedNumbers: number;
  revenue: number;
  recentUsers?: User[];
  otpByCountry?: AdminStatsOtpByCountryItem[];
}

export type ListOtpLogsParams = {
  limit?: number;
  offset?: number;
};

export type ListUsersParams = {
  limit?: number;
  offset?: number;
};