# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatOTP - WhatsApp OTP Verification SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Plans
...
  /api-keys:
    get:
      operationId: listApiKeys
      tags: [api-keys]
      summary: List user API keys
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [api-keys]
      summary: Create a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyWithSecret"
  /api-keys/{id}/revoke:
    post:
      operationId: revokeApiKey
      tags: [api-keys]
      summary: Revoke an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"

  # WhatsApp Numbers (Admin)
...
  /otp/send:
    post:
      operationId: sendOtp
      tags: [otp]
      summary: Send OTP via WhatsApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpSendInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpSendResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /otp/verify:
    post:
      operationId: verifyOtp
      tags: [otp]
      summary: Verify OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyInput"
      responses:
        "200":
          description: OTP verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResponse"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # OTP Logs (for user dashboard)
  /otp/logs:
    get:
      operationId: listOtpLogs
      tags: [otp]
      summary: List OTP request logs for current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OTP logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OtpLog"

  # Payments (Paystack)
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWithSubscription"
  /admin/users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user details (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithSubscription"
  /admin/users/{id}/suspend:
    post:
      operationId: suspendUser
      tags: [users]
      summary: Suspend a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/assign-plan:
    post:
      operationId: adminAssignPlan
      tags: [users]
      summary: Assign a plan to a user (admin override)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignPlanInput"
      responses:
        "200":
          description: Plan assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Auth
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
          minLength: 1
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        suspended:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserWithSubscription:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        suspended:
          type: boolean
        createdAt:
          type: string
          format: date-time
        subscription:
          $ref: "#/components/schemas/Subscription"
        otpCount:
          type: integer

    # Plans
    Plan:
      type: object
      required: [id, name, period, price, currency, otpLimit, features]
      properties:
        id:
          type: integer
        name:
          type: string
        period:
          type: string
          enum: [weekly, monthly, yearly]
        price:
          type: number
        currency:
          type: string
        otpLimit:
          type: integer
        features:
          type: array
          items:
            type: string
        popular:
          type: boolean
        allowCustomNumber:
          type: boolean
        allowUsaNumbers:
          type: boolean

    # Subscriptions
    Subscription:
      type: object
      required: [id, userId, planId, status, startDate, endDate]
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        plan:
          $ref: "#/components/schemas/Plan"
        status:
          type: string
          enum: [active, cancelled, expired]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        otpUsed:
          type: integer
...
    ApiKey:
      type: object
      required: [id, name, keyPrefix, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        keyPrefix:
          type: string
        status:
          type: string
          enum: [active, revoked]
        lastUsedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        requestCount:
          type: integer
    ApiKeyWithSecret:
      type: object
      required: [id, name, keyPrefix, status, createdAt, secret]
      properties:
        id:
          type: integer
        name:
          type: string
        keyPrefix:
          type: string
        status:
          type: string
          enum: [active, revoked]
        lastUsedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        requestCount:
          type: integer
        secret:
          type: string
    ApiKeyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    # Admin plan assignment
    AssignPlanInput:
      type: object
      required: [planId]
      properties:
        planId:
          type: integer

    # WhatsApp display name change
...
    OtpSendInput:
      type: object
      required: [phoneNumber]
      properties:
        phoneNumber:
          type: string
          minLength: 7
        template:
          type: string
    OtpSendResponse:
      type: object
      required: [requestId, expiresAt]
      properties:
        requestId:
          type: string
        expiresAt:
          type: string
          format: date-time
        message:
          type: string
    OtpVerifyInput:
      type: object
      required: [requestId, code]
      properties:
        requestId:
          type: string
        code:
          type: string
          minLength: 4
    OtpVerifyResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        message:
          type: string
    OtpLog:
      type: object
      required: [id, phoneNumber, status, createdAt]
      properties:
        id:
          type: integer
        phoneNumber:
          type: string
        status:
          type: string
          enum: [pending, sent, verified, failed, expired]
        requestId:
          type: string
        country:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # Dashboard
...