# lib/api-zod/src/generated/types/apiKeyWithSecret.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatOTP - WhatsApp OTP Verification SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKeyWithSecretStatus } from "./apiKeyWithSecretStatus";

export interface ApiKeyWithSecret {
  id: number;
  name: string;
  keyPrefix: string;
  status: ApiKeyWithSecretStatus;
  /** @nullable */
  lastUsedAt?: Date | null;
  createdAt: Date;
  requestCount?: number;
  secret: string;
}