# lib/api-zod/src/generated/types/assignPlanInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatOTP - WhatsApp OTP Verification SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignPlanInput {
  planId: number;
}