# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatOTP - WhatsApp OTP Verification SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanPeriod } from "./planPeriod";

export interface Plan {
  id: number;
  name: string;
  period: PlanPeriod;
  price: number;
  currency: string;
  otpLimit: number;
  features: string[];
  popular?: boolean;
  allowCustomNumber?: boolean;
  allowUsaNumbers?: boolean;
}